Isolate gang scheduling tests
